Updated checksum (BSD)
Due to the introducing of incremental (differential) backup, the backup zip files are updated with their logs for tracking file checksum differences, which is resulting in two identical backups to not have the same checksum (since the log file names are different). In this update, we are capturing and logging the archive file checksum before adding the log to the archive. New archives will reflect the newly expected checksum whereas old archives that were captured before this change will reflect the actual archive checksum with the log.